+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Minor: User popup will now automatically display messages as they are received
 - Minor: Changed the English in two rate-limited system messages (#1878)
 - Minor: Added image for streamer mode in the user popup icon.
+- Minor: Added vip and unvip buttons.
 - Bugfix: Fix bug preventing users from setting the highlight color of the second entry in the "User" highlights tab (#1898)
 - Bugfix: Fix bug where the "check user follow state" event could trigger a network request requesting the user to follow or unfollow a user. By itself its quite harmless as it just repeats to Twitch the same follow state we had, so no follows should have been lost by this but it meant there was a rogue network request that was fired that could cause a crash (#1906)
 - Bugfix: /usercard command will now respect the "Automatically close user popup" setting (#1918)